pyproject.toml
[project]
name = "veeam-metrics-collector"
version = "1.0.1"
description = "Veeam Backup Metrics collector"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "datetime>=5.5",
    "influxdb-client>=1.48.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentry-sdk>=2.22.0",
]
